[helper Logging] Update: size_t instead of int for the logging buffer size and access.

+++ SofaKernel/framework/sofa/helper/logging/LoggingMessageHandler.cpp
@@ -46,19 +46,19 @@ void LoggingMessageHandler::reset()
     m_messages.clear() ;
 }
 
-int LoggingMessageHandler::activate()
+size_t LoggingMessageHandler::activate()
 {
     assert(m_activationCount>=0) ;
     m_activationCount++;
     return m_messages.size() ;
 }
 
-int LoggingMessageHandler::deactivate()
+size_t LoggingMessageHandler::deactivate()
 {
     assert(m_activationCount>0) ;
     m_activationCount--;
 
-    int size = m_messages.size();
+    size_t size = m_messages.size();
 
     if(m_activationCount<=0)
         m_messages.clear() ;
@@ -81,12 +81,12 @@ LoggingMessageHandler& MainLoggingMessageHandler::getInstance()
     return s_instance;
 }
 
-int MainLoggingMessageHandler::activate()
+size_t MainLoggingMessageHandler::activate()
 {
     return getInstance().activate() ;
 }
 
-int MainLoggingMessageHandler::deactivate()
+size_t MainLoggingMessageHandler::deactivate()
 {
     return getInstance().deactivate() ;
 }
